feat: Add diagnostic information to the device
This encapsulated diagnostic information entirely within the device since
the device has all of the local information such as traits, device
information, product information, etc. The motivation is to facilitate
local testing rather than testing via Home Assistant and dealing with
mocks and it could be helpful for other users of the library as well.