feat: Simplify device manager creation
Remove the need to pass in a home data api by taking all information up
 front to create the roborock web API in the device manager creation
interface.

This creates a new Web Api wrapper that will also be used for getting
routine information in a followup PR, which was the initial motivation
behind this change.

Overall this will allow additional code simplification in callers when
onboarding the new API, completely removing the need to continue using
the Web API directly for devices.